requirements.txt
asgiref==3.8.1
Django==5.2
django-role-permissions==3.2.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
pypi==2.1
sqlparse==0.5.3
tzdata==2025.2
